- switch to mongodb

+++ src/main/java/com/nprogramming/interviewing/CandidateController.java
@@ -3,6 +3,8 @@ package com.nprogramming.interviewing;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigInteger;
+
 @RestController
 @RequestMapping("/candidates")
 public class CandidateController {
@@ -25,13 +27,13 @@ public class CandidateController {
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.PUT)
-    public Candidate update(@RequestBody Candidate updatedItem, @PathVariable Long id) {
+    public Candidate update(@RequestBody Candidate updatedItem, @PathVariable BigInteger id) {
         updatedItem.setId(id);
         return repository.save(updatedItem);
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
-    public void delete(@PathVariable Long id) {
+    public void delete(@PathVariable BigInteger id) {
         repository.delete(id);
     }
 }
